fix: Onboarding naming and typesetting
